+++ .github/workflows/deploy-dev.yaml
@@ -1,0 +1,47 @@
+name: deploy-dev
+
+on:
+  push:
+    branches: ['master']
+
+env:
+  AWS_DEFAULT_REGION: 'ap-northeast-2'
+  AWS_HOST_ZONE: ${{ secrets.AWS_HOST_ZONE }}
+  AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+  AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+  DOMAIN_NAME: ${{ secrets.DEV_DOMAIN_NAME }}
+
+jobs:
+  deploy-dev:
+    name: deploy to development environment
+    timeout-minutes: 15
+    runs-on: ubuntu-latest
+    # To use Remote Caching, uncomment the next lines and follow the steps below.
+    # env:
+    #  TURBO_TOKEN: ${{ secrets.TURBO_TOKEN }}
+    #  TURBO_TEAM: ${{ vars.TURBO_TEAM }}
+
+    steps:
+      - name: Check out code
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 2
+
+      - uses: pnpm/action-setup@v2.0.1
+        with:
+          version: 8.6.3
+
+      - name: Setup Node.js environment
+        uses: actions/setup-node@v3
+        with:
+          node-version: 18
+          cache: 'pnpm'
+
+      - name: Install dependencies
+        run: pnpm install
+
+      - name: Build
+        run: pnpm build
+
+      - name: Test
+        run: pnpm test
