+++ packages/service-backend/package.json
@@ -6,14 +6,14 @@
   "license": "UNLICENSED",
   "main": "dist/index.js",
   "scripts": {
-    "build": "tsc -p tsconfig.build.json --noEmit",
+    "build": "tsc -p tsconfig.build.json --noEmit && npx webpack",
     "test": "NODE_ENV=test jest",
     "lint": "eslint . --ext .js,.ts",
     "dev": "ts-node src/entry.local.ts",
     "test:debug": "node --inspect-brk ./node_modules/jest/bin/jest.js --runInBand",
     "start:debug": "node --require ts-node/register --inspect-brk src/entry.local.ts",
     "log:show": "aws logs tail /aws/lambda/backend-dev --since 60m",
-    "deploy": "make all",
+    "deploy:dev": "STAGE=dev make all",
     "prepare": "ts-patch install"
   },
   "author": {
