+++ packages/service-backend/src/AppModule.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { DatabaseModule } from './core/database';
+import { MetricModule } from './core/metrics/MetricModule';
 import { AuthModule } from './modules/auth/AuthModule';
 import { UserModule } from './modules/user/UserModule';
 
 @Module({
-  imports: [DatabaseModule, UserModule, AuthModule],
+  imports: [MetricModule, DatabaseModule, UserModule, AuthModule],
 })
 export class AppModule {}
