+++ packages/service-backend/src/core/metrics/MetricController.ts
@@ -1,0 +1,12 @@
+import { Controller, Get } from '@nestjs/common';
+import { getSourceVersion } from '../env';
+
+@Controller('metric')
+export class MetricController {
+  @Get('/health')
+  async health() {
+    return {
+      sha: getSourceVersion(),
+    };
+  }
+}
