+++ packages/service-backend/src/core/metrics/MetricModule.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { MetricController } from './MetricController';
+
+@Module({
+  controllers: [MetricController],
+})
+export class MetricModule {}
