+++ packages/service-backend/src/modules/card/CardController.ts
@@ -4,16 +4,15 @@ import { Card } from '../../core/database';
 
 @Controller('cards')
 export class CardController {
-    constructor(
-        private readonly cardService: CardService
-    ) {}
-    
-    @Get()
-    async getTestCard(){
-        return "card";
-    }
-    @Post()
-    async onCreateCard(@Body() cardDto: Omit<Card, 'createdAt' | 'updatedAt' | 'deletedAt'>){
-        return this.cardService.create(cardDto);
-    }
+  constructor(private readonly cardService: CardService) {}
+
+  @Get()
+  async getTestCard() {
+    return 'card';
+  }
+
+  @Post()
+  async onCreateCard(@Body() cardDto: Omit<Card, 'createdAt' | 'updatedAt' | 'deletedAt'>) {
+    return this.cardService.create(cardDto);
+  }
 }
